move into /cashu/animated-qr-codes
