+++ app/lib/cashu/animated-qr-code/animated-qr-code.tsx
